출석 정보 가져오기 구현 (#1015)
* AttendanceCodeDialog 구현 (#852)

* Attendance screen compose setting (#676)

* feat: implement NewAttendanceActivity

* feat: implement NewAttendanceViewModel

* chore: add compose-lifecycle dependency

* feat: define AttendanceAction

* feat: implement screens

* feat: use SoptTheme in designsystem

* chore: data class -> class로 변경

* chore: 람다 프로퍼티 이름 명시

* chore: SoptTheme darkTheme 기본값 사용

* chore: 필요없는 함수 제거

* chore: 구현 안 된 함수에 TODO 삽입

* chore: 동작하지 않는 Preview 제거

* chore: AttendanceAction 내 뷰모델 참조 제거

* chore: code format 변경

* chore: make stamp design system internal

* chore: extract string resource

* feat: implement AttendanceCodeCard

* feat: implement AttendanceCodeCardList

* chore: change logic

* feat: implement AttendanceCodeDialog

* feat: implement attendance button

* chore: string resource 추출

* chore: change parameter List to ImmutableList

* chore: reformat codee

* feat: implement domain models

* feat: implement Repository

* chore: remove unnecessary code

* chore: change attendanceScore to Float

* chore: reflect type name changed

* chore: remove totalCount from parameters

* feat: add action onClickRefresh

* feat: implement bindDefaultAttendanceRepository

* chore: add background at route

* chore: change button radius

* chore: reformat code

* feat: implement Domain Model to UiModel

* chore: rename AttendanceInfo to Attendance

* chore: rename SessionInfo

* feat: format LocalDateTime

* chore: use DateTimeFormatter

* chore: use getOrNull

* chore: don't use it

* feat: implement AttendanceMapper

* chore: remove duplicate code

* chore: change to Extension

* chore: change parameter name

* chore: change code clear

* chore: make code clear

* chore: separate package

* chore: change function name

* feat: implement FinalAttendanceTest

* chore: remove duplicate tests

+++ app/src/main/java/org/sopt/official/feature/attendance/compose/component/AttendanceHistoryCard.kt
@@ -24,7 +24,7 @@ import org.sopt.official.feature.attendance.model.AttendanceUiState.Success.Atte
 @Composable
 fun AttendanceHistoryCard(
     userTitle: String,
-    attendanceScore: Int,
+    attendanceScore: Float,
     totalAttendanceResult: Map<AttendanceResultType, Int>,
     attendanceHistoryList: ImmutableList<AttendanceHistory>,
     scrollState: ScrollState,
@@ -72,7 +72,7 @@ private fun AttendanceHistoryCardPreview() {
     SoptTheme {
         AttendanceHistoryCard(
             userTitle = "32기 디자인파트 김솝트",
-            attendanceScore = 1,
+            attendanceScore = 1f,
             totalAttendanceResult = mapOf(
                 Pair(AttendanceResultType.ALL, 16),
                 Pair(AttendanceResultType.PRESENT, 5),
